chore: temp commit
Note: this needs to be reset back and reviewed. This temp commit covers
all of the last changes to staging for the reworking of metadata, but
we need to re-run some of these scripts with new changes merging over
from master to staging.

Committing temporarily so I don't lose these files.